requirements.txt
mcp>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.33.0
nest_asyncio>=1.5.8 